requirements.txt
# Data Collection & Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3
webdriver-manager==4.0.1

# ML/AI - Embeddings & Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.1

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Frontend
streamlit==1.28.0

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
tenacity==8.2.3

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0
ipykernel==6.27.1

# Optional: For enhanced features
# openai==1.3.7  # If using OpenAI embeddings
# google-generativeai==0.3.1  # If using Gemini
